+++ src/components/controlled/transactions-table/tests/transactions-table.spec.tsx
@@ -3,6 +3,7 @@ import { newSpecPage } from '@stencil/core/testing';
 
 import { TransactionsTable } from '../transactions-table';
 import type { TransactionRowType } from '../transactions-table.type';
+import { IconNamesEnum } from 'common/Icon/icon.types';
 
 describe('TransactionsTable', () => {
   const mockTransactions: TransactionRowType[] = [
@@ -10,7 +11,7 @@ describe('TransactionsTable', () => {
       age: { timeAgo: '5 minutes ago', tooltip: 'Feb 7, 2025, 4:55 PM' },
       direction: 'in',
       method: { name: 'transfer', actionDescription: 'Token transfer' },
-      iconInfo: { icon: 'check', tooltip: 'Successful transaction' },
+      iconInfo: { icon: IconNamesEnum.check, tooltip: 'Successful transaction' },
       link: '/transactions/hash1',
       receiver: {
         address: 'erd1qqqqqqqqqqqqqpgqp699jngundfqw07d8jzkepucvpzush6k3wvqyc44rx',
@@ -54,7 +55,7 @@ describe('TransactionsTable', () => {
       age: { timeAgo: '10 minutes ago', tooltip: 'Feb 7, 2025, 4:50 PM' },
       direction: 'out',
       method: { name: 'esdt_transfer', actionDescription: 'ESDT Token transfer' },
-      iconInfo: { icon: 'check', tooltip: 'Successful transaction' },
+      iconInfo: { icon: IconNamesEnum.check, tooltip: 'Successful transaction' },
       link: '/transactions/hash2',
       receiver: {
         address: 'erd1qyu5wthldzr8wx5c9ucg8kjagg0jfs53s8nr3zpz3hypefsdd8ssycr6th',
@@ -126,20 +127,13 @@ describe('TransactionsTable', () => {
     const rows = page.root.querySelectorAll('tbody tr');
     expect(rows.length).toBe(2);
 
-    rows.forEach((row, index) => {
-      expect(row.querySelector('mvx-transaction-hash')).toBeTruthy();
-      expect(row.querySelector('mvx-transaction-age')).toBeTruthy();
-      expect(row.querySelector('mvx-transaction-shards')).toBeTruthy();
-      expect(row.querySelector('mvx-transaction-account[scope="sender"]')).toBeTruthy();
-      expect(row.querySelector('mvx-transaction-account[scope="receiver"]')).toBeTruthy();
-      expect(row.querySelector('mvx-transaction-method')).toBeTruthy();
-      expect(row.querySelector('mvx-transaction-value')).toBeTruthy();
+    const cells = rows[0].querySelectorAll('td');
+    expect(cells.length).toBe(7);
 
-      // Check some specific values
-      expect(row.querySelector('mvx-transaction-age').getAttribute('age')).toBe(mockTransactions[index].age.timeAgo);
-      expect(row.querySelector('mvx-transaction-method').getAttribute('method')).toBe(
-        mockTransactions[index].method.name,
-      );
-    });
+    // Check some specific values
+    const ageCell = cells[1];
+    expect(ageCell.textContent).toContain('5 minutes ago');
+    const methodCell = cells[5];
+    expect(methodCell.textContent).toContain('transfer');
   });
 });