+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[0.0.36](https://github.com/multiversx/mx-sdk-dapp-ui/pull/251)] - 2025-10-15
+
+- [Updated icons type](https://github.com/multiversx/mx-sdk-dapp-ui/pull/250)
+
+- [Added the Storybook as a live server deployment](https://github.com/multiversx/mx-sdk-dapp-ui/pull/248)
+
+- [Refactor components in transactions table](https://github.com/multiversx/mx-sdk-dapp-ui/pull/243)
+
 ## [[0.0.35](https://github.com/multiversx/mx-sdk-dapp-ui/pull/246)] - 2025-10-10
 
 - [Added passkey provider](https://github.com/multiversx/mx-sdk-dapp-ui/pull/242)